Add READ_PHONE_STATE back to pregranted phone permissions
When READ_PHONE_STATE was moved from a runtime permission to an install
permission it was removed from the default phone permissions grant. Now
that READ_PHONE_STATE has been reverted back to a runtime permission it
needs to be added back to the list of phone permissions that are
pregranted.

Bug: 154572075
Bug: 154567729
Test: Verified com.google.shared.uid was granted READ_PHONE_STATE
Change-Id: I0e0727ed92cbf881d89d56de4e509fb1a9073556
(cherry picked from commit 5a6845480efc780e3aa1129b6ef84ae7867b2539)