Support split-screen modes on freeform displays
This was originally not allowed to prevent mistakes while
transitioning to new split-screen and didn't pose a problem
because Tiles didn't hit this check. Now that the transition
is complete, this is not needed. Also, now that everything
is Tasks, this check prevents creating the root surfaces
if the display is currently freeform. This doesn't make
sense because a display can change to freeform after adding
the root Tasks and it'd be fine.

Bug: 151881448
Test: Create root split task on freeform display (eg. via ARC)
Change-Id: I90a1ba67d2c6446e709c705458c4f6c33b4007cc