Only notify when admin enables location for the user
Only notify the user when the admin enables location services on their
device, not on every state change.

Prior to this change the user would be notified when the admin disables
location services, with a misleading notification indicating apps now
have location available.

Tested by installing TestDPC, setting it as device owner, then toggling
location on and off and observing user is notified only on turning
location on.
Test both legacy and new method flows.

Bug: 156602188
Test: Manual using TestDPC (see above)
Change-Id: Ia757802bd7da1afeecfd8071598b692c69860c66