Updated the paddings of quick settings to match notifications
Now all content is nicely aligned on the same line like the
notifications. Previously this was a mix of statusbar alignment
and some other alignment, that sometimes worked out to actually
be at the same location. The setting of these paddings
is now streamlined in code.
This also ensures that the buttons themselves now nicely align
on the left side of the view instead of being 1dp off.
This also fixes a bug where the paddings were off due to the
last media refactorings on Pixel 2 or non-rounded displays.

Fixes: 156452277
Bug: 157258041
Test: test layout of Qs with and without media notification, with various cutouts.
Change-Id: Id547f9a99577886522e41bdb7e3aedac47db9a9e