Remove unused restore-mode logic in ActivityStack
This was added when split-screen was implemented in
a flat manner (and thus Wm had to remember what mode
the stack was in before going to split). Now that
split-screen is based on inheritence, this is basically
dead code.

Additionally, now that we use task-organizers for
split, this actually gets in the way because it can
end-up doing something the organizer didn't ask for.

Bug: 151881448
Test: Existing tests pass.
Change-Id: I9fecdc704499c09c2e2924081b2e323d9b85ae0a