Keep secondary prox registered when covered.
We were pinging the secondary proximity sensor every 5 seconds to see
if the "near" state had changed. However, after some amount of time,
this 5 second delay would be extended as the cpu would stop waking up,
presumably for power saving reasons.

Instead, we now keep the secondary proxmity sensor registered when
the phone is covered. This means that the phone will immediately wake
up when uncovered, as expected.

The ThresholdSensorImpl now reports a change from below to above when
the latch value is equaled, rather surpassed. This ensures that some,
coarse grained sensors will still trigger when a small change in their
value is detected. Prior to this, a larger change was needed, meaning
that uncovering a phone in a dark room could leave the screen off.

Fixes: 157630609
Test: atest SystemUITests && manual
Change-Id: I4b4e936b19045d38445bf6ab92df2bb6e3d90a28